requirements.txt
asgiref==3.11.0
black==26.1.0
click==8.3.1
Django==6.0.1
djangorestframework==3.16.1
iniconfig==2.3.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.5.1
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.11.1
pytokens==0.4.1
ruff==0.14.14
sqlparse==0.5.5
